ecommerce_project/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
multidict==6.0.4
openai==0.28.0
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
